Phase 7B Session 5: Complete warning elimination in Src/infoG.c
Prompt: Please start on infoG.c

Work Completed:
- Fixed 3 format specifier warnings in infoG.c (3 → 0 warnings)
- Warning 1: Line 173 - Cast strlen() result to int for printf width specifier in nation materials display
- Warning 2: Line 214 - Cast strlen() result to int for printf width specifier in city materials display
- Warning 3: Line 973 - Changed %d to %ld format specifier for nation score (long int)
- All fixes preserve original formatting behavior while ensuring type safety
- Clean compilation achieved with zero warnings

+++ Src/infoG.c
@@ -170,7 +170,7 @@ disp_city_info PARM_2 (int, xpos, int, line)
     for (i = 0; i < MTRLS_NUMBER; i++) {
       mvprintw(line + i, xpos + CINFO_WIDTH2, "%s:%*.0lf",
 	       mtrls_info[i].name,
-	       18 - strlen(mtrls_info[i].name),
+	       (int)(18 - strlen(mtrls_info[i].name)),
 	       (double) ntn_ptr->mtrls[i]);
     }
 
@@ -211,7 +211,7 @@ disp_city_info PARM_2 (int, xpos, int, line)
     for (i = 0; i < MTRLS_NUMBER; i++) {
       mvprintw(line + i, xpos + CINFO_WIDTH2, "%s:%*.0lf",
 	       mtrls_info[i].name,
-	       18 - strlen(mtrls_info[i].name),
+	       (int)(18 - strlen(mtrls_info[i].name)),
 	       (double) city_Iptr->c_mtrls[i]);
     }
 
@@ -970,7 +970,7 @@ disp_info PARM_3 (int, x, int, y, Infotype, choice)
       mvaddstr( y++, x, ". . .");
       mvaddstr( y++, x, ". . .");
     } else {
-      mvprintw( y++, x, "%d", ntn_Iptr->score );
+      mvprintw( y++, x, "%ld", ntn_Iptr->score );
       mvprintw( y++, x, "%.0f", (double) ntn_Iptr->mtrls[MTRLS_TALONS] );
       mvprintw( y++, x, "%.0f", (double) ntn_Iptr->tmil );
       mvprintw( y++, x, "%.0f", (double) ntn_Iptr->tciv );
